test(csv-export/test-helper): moved test helpers one level up
#236
