+++ tests/exports/fundingrequest_csv/test_invoice_mappers.py
@@ -14,7 +14,7 @@ from coda.apps.exports.services.fundingrequest_csv.mappers import (
 
 from coda.domain.finance.invoice_positions import ContractItem
 from tests import modelfactory
-from tests.exports.fundingrequest_csv.helpers import (
+from tests.exports.helpers import (
     create_funding_request,
     create_invoice_with_funding_assignments,
     create_invoice_with_currency_conversion,
